cdp_support_chatbot/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
scikit-learn==1.3.0
Flask==2.2.3
python-dotenv==1.0.0
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.0

# Development Tools
pytest==7.3.1
black==23.3.0
flake8==6.0.0